Addressed changes requested
- Made actions/bindings pub(crate)
- Moved test to unit tests
- Updated examples
- Removed unnecessary imports
